factor out OptionColumn into a component
