+++ model/decision.ts
@@ -1,0 +1,8 @@
+export interface Option {
+  pros: string[];
+  cons: string[];
+}
+
+type Decision = Option[];
+
+export default Decision;
